mediacapture-streams/MediaStreamTrack-init.https.html

<!doctype html>
<title>getUserMedia({video: true}) creates a stream with a properly initialized video track</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#idl-def-MediaStreamTrack">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#life-cycle-and-media-flow">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-MediaStreamTrack-kind">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-MediaStreamTrack-enabled">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-MediaStreamTrack-readyState">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<p class="instructions">When prompted, accept to share your video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that the video track of MediaStream
object returned by the success callback in getUserMedia is correctly initialized.</p>

<script>

promise_test(async t => {
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  const videoTracks = stream.getVideoTracks();
  assert_equals(videoTracks.length, 1, "There is exactly one video track in the media stream");
  const track = videoTracks[0];
  t.add_cleanup(() => track.stop());
  assert_equals(track.readyState, "live", "The track object is in live state");
  assert_equals(track.kind, "video", "The track object is of video kind");
  assert_true(track.enabled, "The track object is enabed");
});

</script>
